Actual controlled movement

+++ controls/include/motor_control.h
@@ -1,3 +1,3 @@
 void motor_init(void);
 void motor_control(int, int, double);
-void move_bot(double, double, int);
+void move_bot(double, double, double);
